feat: add session apis controller

+++ backend/src/main/java/com/gisma/mentorship_network/service/MentorshipSessionService.java
@@ -1,10 +1,15 @@
 package com.gisma.mentorship_network.service;
 
 import com.gisma.mentorship_network.model.MentorshipSession;
+import com.gisma.mentorship_network.model.SessionStatus;
 import com.gisma.mentorship_network.repository.MentorshipSessionRepository;
 import jakarta.persistence.EntityNotFoundException;
+import jakarta.validation.constraints.NotBlank;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
+import java.util.Date;
 import java.util.List;
 
 @Service
@@ -21,22 +26,43 @@ public class MentorshipSessionService {
 
     public MentorshipSession getMentorshipSessionById(Long id) {
         return mentorshipSessionRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("Mentorship session not found"));
+                .orElseThrow(() -> new ResponseStatusException(
+                        HttpStatus.BAD_REQUEST, "Mentorship session not found"));
     }
 
-    public MentorshipSession createMentorshipSession(MentorshipSession mentorshipSession) {
-        return mentorshipSessionRepository.save(mentorshipSession);
+    public record MentorshipSessionRequest(
+            @NotBlank(message = "Mentorship match is required")
+            Long mentorshipMatchId,
+
+            @NotBlank(message = "Scheduled date is required")
+            Date scheduledDate
+            ) {}
+
+    public MentorshipSession createMentorshipSession(MentorshipSessionRequest request) {
+        MentorshipSession newMentorshipSession = new MentorshipSession();
+        newMentorshipSession.setScheduledDate(request.scheduledDate);
+        newMentorshipSession.setMentorshipMatchId(request.mentorshipMatchId);
+        return mentorshipSessionRepository.save(newMentorshipSession);
     }
 
-    public MentorshipSession updateMentorshipSession(Long id, MentorshipSession mentorshipSession) {
+    public record UpdateMentorshipSessionRequest(
+      Date scheduledDate, SessionStatus status, String mentorNotes
+    ) {}
+
+    public MentorshipSession updateMentorshipSession(Long id, UpdateMentorshipSessionRequest request) {
         MentorshipSession existingSession = getMentorshipSessionById(id);
-        existingSession.setStatus(mentorshipSession.getStatus());
-        existingSession.setScheduledDate(mentorshipSession.getScheduledDate());
-        existingSession.setMentorNotes(mentorshipSession.getMentorNotes());
+        existingSession.setStatus(request.status);
+        existingSession.setScheduledDate(request.scheduledDate);
+        existingSession.setMentorNotes(request.mentorNotes);
         return mentorshipSessionRepository.save(existingSession);
     }
 
     public void deleteMentorshipSession(Long id) {
+        if(!mentorshipSessionRepository.existsById(id)) {
+            throw new ResponseStatusException(
+                    HttpStatus.NOT_FOUND, "Session with ID " + id + " not found.");
+        }
+
         mentorshipSessionRepository.deleteById(id);
     }
 }
